+++ src/main/java/io/supertokens/webserver/Webserver.java
@@ -151,6 +151,8 @@ public class Webserver extends ResourceDistributor.SingletonResource {
         addAPI(new UserAPI(main));
         addAPI(new GenerateEmailVerificationTokenAPI(main));
         addAPI(new VerifyEmailAPI(main));
+        addAPI(new UsersAPI(main));
+        addAPI(new UsersCountAPI(main));
     }
 
     public void addAPI(WebserverAPI api) {