+++ src/main/java/io/supertokens/webserver/api/emailpassword/UserAPI.java
@@ -76,7 +76,11 @@ public class UserAPI extends WebserverAPI {
             } else {
                 JsonObject result = new JsonObject();
                 result.addProperty("status", "OK");
-                result.add("user", new JsonParser().parse(new Gson().toJson(user)).getAsJsonObject());
+                JsonObject userJson = new JsonParser().parse(new Gson().toJson(user)).getAsJsonObject();
+                if (super.getVersionFromRequest(req).equals("2.4")) {
+                    userJson.remove("timeJoined");
+                }
+                result.add("user", userJson);
                 super.sendJsonResponse(200, result, resp);
             }