+++ src/test/java/io/supertokens/test/GeneratePasswordResetTokenAPITest2_4.java
@@ -121,7 +121,7 @@ public class GeneratePasswordResetTokenAPITest2_4 {
             return;
         }
 
-        JsonObject signUpResponse = Utils.signUpRequest(process, "random@gmail.com", "validPass123");
+        JsonObject signUpResponse = Utils.signUpRequest_2_4(process, "random@gmail.com", "validPass123");
         assertEquals(signUpResponse.get("status").getAsString(), "OK");
         assertEquals(signUpResponse.entrySet().size(), 2);
         String userId = signUpResponse.getAsJsonObject("user").get("id").getAsString();